series/requirements/base.txt
pydicom==2.4.4
pynetdicom==2.1.0
aiohttp==3.9.5
python-dotenv==1.0.1